# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Key Service",
    "description": "Manages API keys and BYOK (Bring Your Own Key) credentials for organizations. Handles key generation, validation, encryption, and secure storage.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "serviceKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Service-to-service key (KEY_SERVICE_API_KEY)"
      }
    },
    "schemas": {
...
      "ValidateKeyResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "key"
        ]
      },
      "ListUserAuthKeysResponse": {
        "type": "object",
        "properties": {
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserAuthKeyItem"
            }
          }
        },
        "required": [
          "keys"
        ]
      },
      "UserAuthKeyItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "keyPrefix": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "orgId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "keyPrefix",
          "name",
          "orgId",
          "userId",
          "createdBy",
          "createdAt",
          "lastUsedAt"
        ]
      },
      "CreateUserAuthKeyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "orgId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "key",
          "name",
          "orgId",
          "userId",
          "createdBy",
          "createdAt"
        ]
      },
      "CreateUserAuthKeyRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "minLength": 1
          },
          "createdBy": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "userId",
          "createdBy",
          "name"
        ]
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "SessionKeyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "type": "string"
          },
          "keyPrefix": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "key",
          "keyPrefix",
          "name"
        ]
      },
      "ListOrgKeysResponse": {
        "type": "object",
        "properties": {
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrgKeyItem"
            }
          }
        },
        "required": [
          "keys"
        ]
      },
      "OrgKeyItem": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "maskedKey": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "provider",
          "maskedKey",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateOrgKeyResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "maskedKey": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "maskedKey",
          "message"
        ]
      },
      "CreateOrgKeyRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "minLength": 1
          },
          "apiKey": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "provider",
          "apiKey"
        ]
      },
      "DeleteOrgKeyResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "message"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/validate/keys/{provider}": {
      "get": {
        "summary": "Get decrypted org key for a provider",
        "description": "Pass the user's API key as ?key= query parameter. Requires X-Caller-Service, X-Caller-Method, and X-Caller-Path headers to identify the calling endpoint. These are used to build the provider requirements registry.",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "provider",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "The API key to validate (distrib.usr_*)",
              "example": "distrib.usr_abc123"
            },
            "required": true,
            "description": "The API key to validate (distrib.usr_*)",
            "name": "key",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Name of the calling service",
              "example": "apollo"
            },
            "required": true,
            "description": "Name of the calling service",
            "name": "x-caller-service",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "HTTP method of the caller's endpoint",
              "example": "POST"
            },
            "required": true,
            "description": "HTTP method of the caller's endpoint",
            "name": "x-caller-method",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Path of the caller's endpoint",
              "example": "/leads/search"
            },
            "required": true,
            "description": "Path of the caller's endpoint",
            "name": "x-caller-path",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Decrypted key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing key parameter or required caller headers"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Key not configured"
          }
        }
      }
    },
    "/api-keys": {
      "get": {
        "summary": "List user auth keys for an org",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "userId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "List of user auth keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUserAuthKeysResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user auth key",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserAuthKeyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User auth key created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserAuthKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/api-keys/{id}": {
      "delete": {
        "summary": "Delete a user auth key",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "User auth key deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Key not found"
          }
        }
      }
    },
    "/api-keys/session": {
      "post": {
        "summary": "Get or create a default user auth key for the org+user",
        "description": "Uses orgId and userId from identity headers (x-org-id, x-user-id). No request body needed.",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Session key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionKeyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/keys": {
      "get": {
        "summary": "List org keys for an org",
        "description": "Uses orgId from identity header (x-org-id).",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "List of org keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOrgKeysResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add or update an org key",
        "description": "Uses orgId from identity header (x-org-id).",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgKeyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Org key saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrgKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/keys/{provider}": {
      "delete": {
        "summary": "Delete an org key",
        "description": "Uses orgId from identity header (x-org-id).",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "provider",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Org key deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteOrgKeyResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}